+++ src/main/java/com/twilio/rest/messaging/v1/TollfreeVerificationUpdater.java
@@ -66,6 +66,8 @@ public class TollfreeVerificationUpdater extends Updater<TollfreeVerification> {
     private String termsAndConditionsUrl;
     private Boolean ageGatedContent;
     private List<String> optInKeywords;
+    private TollfreeVerification.VettingProvider vettingProvider;
+    private String vettingId;
 
     public TollfreeVerificationUpdater(final String pathSid) {
         this.pathSid = pathSid;
@@ -324,6 +326,18 @@ public class TollfreeVerificationUpdater extends Updater<TollfreeVerification> {
         final String optInKeywords
     ) {
         return setOptInKeywords(Promoter.listOfOne(optInKeywords));
+    }
+
+    public TollfreeVerificationUpdater setVettingProvider(
+        final TollfreeVerification.VettingProvider vettingProvider
+    ) {
+        this.vettingProvider = vettingProvider;
+        return this;
+    }
+
+    public TollfreeVerificationUpdater setVettingId(final String vettingId) {
+        this.vettingId = vettingId;
+        return this;
     }
 
     @Override
@@ -668,6 +682,24 @@ public class TollfreeVerificationUpdater extends Updater<TollfreeVerification> {
                     ParameterType.URLENCODED
                 );
             }
+        }
+
+        if (vettingProvider != null) {
+            Serializer.toString(
+                request,
+                "VettingProvider",
+                vettingProvider,
+                ParameterType.URLENCODED
+            );
+        }
+
+        if (vettingId != null) {
+            Serializer.toString(
+                request,
+                "VettingId",
+                vettingId,
+                ParameterType.URLENCODED
+            );
         }
     }
 }